--- order/.gitignore
@@ -1,4 +1,0 @@
-# package directories
-node_modules
-jspm_packages
-
