+++ order/function.yml
@@ -1,33 +1,33 @@
 getAllOrders:
-    handler: Order_Apis/getAllOrders.getAllOrders
+    handler: Order/getAllOrders.getAllOrders
     events:
       - httpApi:
           path: /getAllOrders
           method: get
 
 createOrder:
-    handler: Order_Apis/createOrder.createOrder
+    handler: Order/createOrder.createOrder
     events:
       - httpApi:
           path: /createOrder
           method: post   
 
 getOrderById:
-    handler: Order_Apis/getOrderById.getOrderById
+    handler: Order/getOrderById.getOrderById
     events:
       - httpApi:
           path: /getOrderById/{id}
           method: get                        
 
 updateOrder:
-    handler: Order_Apis/updateOrder.updateOrder
+    handler: Order/updateOrder.updateOrder
     events:
       - httpApi:
           path: /updateOrder/{id}
           method: put    
 
 deleteOrderById: 
-    handler: Order_Apis/deleteOrderById.deleteOrderById
+    handler: Order/deleteOrderById.deleteOrderById
     events:
       - httpApi:
           path: /deleteOrderById/{id}
