--- order/serverless.yml
@@ -1,53 +1,0 @@
-service: order
-
-frameworkVersion: "3"
-
-plugins:
- - serverless-offline
-
-provider:
-  name: aws
-  runtime: nodejs18.x
-
-functions:
-  createOrder:
-    handler: createOrder.createOrder
-    timeout: 600
-    events:
-      - http:
-          method: POST
-          path: createorder
-
-  getallOrder:
-    handler: getAllOrder.getAllOrders
-    timeout: 600
-    events:
-      - http:
-          method: get
-          path: getallOrder
-
-  getByOrderId:
-    handler: getById.getByOrderId
-    timeout: 600
-    events: 
-      - http:
-          method: get
-          path: getbyid/{orderId}   
-
-  deleteByOrderId:
-    handler: deleteById.deleteByOrderId
-    timeout: 600
-    events:
-      - http:
-          method:  delete
-          path: deletebyid/{orderId}
-
-  update:
-    handler: updateOrder.updateOrder
-    timeout: 600
-    events: 
-      - http: 
-          method: put
-          path: update/{orderId}
-
-
