+++ serverless.yml
@@ -75,7 +75,7 @@ functions:
   - ${file(Users/function.yml)}
   - ${file(Login/function.yml)}
   - ${file(Customer/function.yml)}
-  - ${file(Order_Apis/function.yml)}
+  - ${file(Order/function.yml)}
   - ${file(catalog/function.yml)}
